requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tensorflow==2.15.0
pillow==10.1.0
flwr==1.8.0
numpy==1.24.3
python-multipart==0.0.6
